refac(example): rename multi-prover to mmp
