+++ sdk/src/mmp/common/mod.rs
