+++ sdk/src/mmp/common/types.rs
