+++ sdk/src/mmp/operator/utils.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 use sp1_core::utils::baby_bear_poseidon2::Perm;
 use std::{fs::File, io::Read};
 
-use crate::multi_prover::common::types::ChallengerType;
+use crate::mmp::common::types::ChallengerType;
 
 pub fn read_bin_file_to_vec(mut file: File) -> Result<Vec<u8>> {
     let metadata = file.metadata()?;
